+++ tests/test_occ_site_functions_chebychev.py
@@ -59,8 +59,6 @@ def test_chebychev_occ_fcc_1a(session_shared_datadir):
     functions, clusters = (builder.functions, builder.clusters)
 
     _occ_site_functions = builder.occ_site_functions.copy()
-    for x in _occ_site_functions:
-        x["value"] = x["value"].tolist()
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0, 1, 2, 3]:
@@ -134,8 +132,6 @@ def test_chebychev_occ_fcc_1b(session_shared_datadir):
     functions, clusters = (builder.functions, builder.clusters)
 
     _occ_site_functions = builder.occ_site_functions.copy()
-    for x in _occ_site_functions:
-        x["value"] = x["value"].tolist()
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0, 1, 3]:
@@ -220,8 +216,6 @@ def test_chebychev_occ_fcc_2a(session_shared_datadir):
     functions, clusters = (builder.functions, builder.clusters)
 
     _occ_site_functions = builder.occ_site_functions.copy()
-    for x in _occ_site_functions:
-        x["value"] = x["value"].tolist()
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0]:
@@ -323,8 +317,6 @@ def test_chebychev_occ_fcc_2b(session_shared_datadir):
     functions, clusters = (builder.functions, builder.clusters)
 
     _occ_site_functions = builder.occ_site_functions.copy()
-    for x in _occ_site_functions:
-        x["value"] = x["value"].tolist()
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0]:
@@ -468,8 +460,6 @@ def test_chebychev_occ_fcc_3a(session_shared_datadir):
     functions, clusters = (builder.functions, builder.clusters)
 
     _occ_site_functions = builder.occ_site_functions.copy()
-    for x in _occ_site_functions:
-        x["value"] = x["value"].tolist()
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0]:
@@ -605,8 +595,6 @@ def test_chebychev_occ_fcc_3b(session_shared_datadir):
     functions, clusters = (builder.functions, builder.clusters)
 
     _occ_site_functions = builder.occ_site_functions.copy()
-    for x in _occ_site_functions:
-        x["value"] = x["value"].tolist()
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0]: